refactor: let a registry's owner read its own keys
Core knew ESMValTool's reference layout only because `_get_parser` dispatched on
registry name, extending a pattern that already hardcoded `pmp-climatology` and `obs4ref`.

- Adds a `key_parser` hook to `RegistryRequest`.
  A registry whose keys core cannot read supplies its own reader.
- Moves the ESMValTool key parser into `climate_ref_esmvaltool`,
  which owns the registry the keys belong to.
  `climate_ref_core.esgf` no longer mentions ESMValTool.
- Moves `frequency_from_mip_table` to `climate_ref_core.cmor`.
  It maps CMIP6 CMOR tables and has nothing to do with ESMValTool,
  so a module named after one provider was the wrong home for it.

The path parsing itself stays in core because the ingest adapter lives in `climate_ref`
and the recipe runner in `climate_ref_esmvaltool`, and neither package can see the other.
