+++ changelog/843.feature.md
@@ -1,4 +1,4 @@
-Added a `RegistryRequest` parser for the ESMValTool reference registry,
-so a test case can fetch the reference data an opt-in diagnostic selects.
-The fetch and ingest sides now parse through the same function,
+Added a `key_parser` to `RegistryRequest`, so the package that owns a registry owns how its keys are read.
+The ESMValTool provider uses it to fetch the reference data an opt-in diagnostic selects.
+The fetch and ingest sides parse through the same function,
 so a request cannot name a facet value that ingest would spell differently.
