+++ packages/climate-ref-core/tests/unit/test_esmvaltool_reference.py
@@ -2,7 +2,6 @@ import pytest
 
 from climate_ref_core.esmvaltool_reference import (
     drs_relative_parts,
-    frequency_from_mip_table,
     parse_reference_path,
     tier_from_segment,
 )
@@ -112,70 +111,6 @@ def test_drs_relative_parts_names_the_project_it_could_not_fit():
 def test_drs_relative_parts_rejects_a_path_with_no_anchor():
     with pytest.raises(ValueError, match="not under a known ESMValTool reference project"):
         drs_relative_parts("/data/somewhere/mystery.nc")
-
-
-class TestFrequencyFromMipTable:
-    """``Amon`` -> ``mon``, for ESMValTool OBS/OBS6 reference data."""
-
-    @pytest.mark.parametrize(
-        "mip_table, expected",
-        [
-            # The common monthly tables across realms.
-            ("Amon", "mon"),
-            ("Omon", "mon"),
-            ("Lmon", "mon"),
-            ("LImon", "mon"),
-            ("SImon", "mon"),
-            ("AERmon", "mon"),
-            ("Emon", "mon"),
-            # Daily.
-            ("day", "day"),
-            ("Oday", "day"),
-            ("SIday", "day"),
-            ("CFday", "day"),
-            # CMIP5-era sea-ice tables (used by ESMValTool OBS data, e.g. OSI-450 sic).
-            ("OImon", "mon"),
-            ("OIday", "day"),
-            # Sub-daily, including the point-sampled variants.
-            ("3hr", "3hr"),
-            ("E3hrPt", "3hrPt"),
-            ("6hrLev", "6hr"),
-            ("6hrPlevPt", "6hrPt"),
-            ("AERhr", "1hr"),
-            # Fixed fields and yearly.
-            ("fx", "fx"),
-            ("Ofx", "fx"),
-            ("Oyr", "yr"),
-            ("IyrGre", "yr"),
-            # The irregular ones the enumeration exists for.
-            ("Oclim", "monC"),
-            ("E1hrClimMon", "1hrCM"),
-            ("Esubhr", "subhrPt"),
-            # Zonal-mean tables keep the frequency of their non-zonal counterpart.
-            ("AERmonZ", "mon"),
-            ("EdayZ", "day"),
-            ("E6hrZ", "6hr"),
-        ],
-    )
-    def test_maps_mip_table_to_frequency(self, mip_table, expected):
-        assert frequency_from_mip_table(mip_table) == expected
-
-    @pytest.mark.parametrize("frequency", ["mon", "day", "fx", "yr", "3hr", "subhrPt"])
-    def test_existing_frequency_passes_through(self, frequency):
-        """``native6`` paths already carry a frequency, so the same call site handles them."""
-        assert frequency_from_mip_table(frequency) == frequency
-
-    @pytest.mark.parametrize("value", ["day", "fx"])
-    def test_names_that_are_both_table_and_frequency_are_idempotent(self, value):
-        assert frequency_from_mip_table(value) == value
-        assert frequency_from_mip_table(frequency_from_mip_table(value)) == value
-
-    @pytest.mark.parametrize("value", ["Amonthly", "", "monthly", "Xmon", "AMON"])
-    def test_unknown_value_raises(self, value):
-        """Failing loudly beats defaulting: a silent fallback would collapse two datasets that
-        differ only by frequency onto one ``instance_id``."""
-        with pytest.raises(ValueError, match="Unknown MIP table or frequency"):
-            frequency_from_mip_table(value)
 
 
 class TestParseReferencePath:
